Add projects heading without grid gap method
It is not part of a grid container so it is not a grid item.
It is simply a regular block level element with grid items.
The projects container despite being a display grid is a block
level grid element, remember. Hence it's sized to content. Others are
grid-items which auto-stretch to cell. Hence give it a min-height or
height. But allow it to get bigger I guess, which I hope is the right
choice.