Fix search icon pic placeholder causing large layout bug
Still not sure why. Removing the icon within the container fixed the
issue. I saw modifying the selector targeting it was not working.
Removing selector targeting all pics fixed it for some reason, though
there is no obvious declaration that should be responsible for the size
issue except maybe the min-height making it grow without bounds.