Trigger cascading enlarged page bug from header item size increases
Here is the reason why: By increasing the content inside the cell,
it resizes as can be seen via dev tools inspection. However since I use
relative fr units first the next section increases by ratio to maintain
the ratio. Furthermore then the articles section increases to keep the
ratio as the header has just increased. The cards automatically grow to
fill the newly available space of the 5fr article section, taking 1fr
inside it, and cards take 1fr each of that. Basically because of this
they expand due to their size being defined by the fr units as the size
grows.

To fix:
- Remove the relationship between them and make header based off another
unit. This will also make header not be resized I think. If its at a
fixed size regardless of content it will not resize, the grid cell
allegedly won't be resized either.
- Prevent content resizing grid cell, and instead allow it to just
overflow. Overflow is not the ideal solution though. To fit everything
consider the units of the header and if they should be fixed or
relative.